+++ src/Timer/GPUTimer.cpp
@@ -9,7 +9,7 @@ GPUTimer::GPUTimer(const Context* context, GPUTimerCreateInfo createInfo) : cont
     timestampPeriod = context->getPhysicalDevice().getProperties().limits.timestampPeriod;
 }
 
-auto GPUTimer::elapsedInNano() -> double {
+auto GPUTimer::elapsedInNano() -> float {
     timestamps.fill(0);
     vk::resultCheck(context->getDevice().getQueryPoolResults(
                         *queryPool, 0, 2,
@@ -18,10 +18,10 @@ auto GPUTimer::elapsedInNano() -> double {
                         sizeof(uint64_t),                      // stride
                         vk::QueryResultFlagBits::e64 | vk::QueryResultFlagBits::eWait),
                     "Failed to get query pool results");
-    return timestampPeriod * static_cast<double>(timestamps[1] - timestamps[0]);
+    return timestampPeriod * static_cast<float>(timestamps[1] - timestamps[0]);
 }
 
-auto GPUTimer::elapsedInMilli() -> double {
-    return elapsedInNano() / 1000000.0;
+auto GPUTimer::elapsedInMilli() -> float {
+    return elapsedInNano() / 1000000.0f;
 }
 }  // namespace rv
